[zhjwang] et

+++ example/passthrough.c
@@ -165,6 +165,13 @@ static int xmp_unlink(const char *path)
 {
 	int res;
 
+	syslog(LOG_ERR, "zhjwang...xmp_unlink...path:%s\n", path);
+	int res_open_cm = strcmp(path, "/home/zhjwang/test/1.et");
+	if (res_open_cm == 0)
+	{
+		return -1;
+	}
+
 	res = unlink(path);
 	if (res == -1)
 		return -errno;
@@ -281,6 +288,7 @@ static int xmp_create(const char *path, mode_t mode,
 {
 	int res;
 
+	syslog(LOG_ERR, "zhjwang...xmp_create...path:%s\n", path);
 	res = open(path, fi->flags, mode);
 	if (res == -1)
 		return -errno;
@@ -295,7 +303,7 @@ static int xmp_open(const char *path, struct fuse_file_info *fi)
 	int res;
 
 	syslog(LOG_ERR, "zhjwang...xmp_open...path:%s\n", path);
-	int res_open_cm = strcmp(path, "/home/zhjwang/test/1.wps");
+	int res_open_cm = strcmp(path, "/home/zhjwang/test/1.et");
 	if (res_open_cm == 0 && write_block_flag == 1)
 	{
 		syslog(LOG_ERR, "zhjwang...xmp_open...return -1\n");
@@ -344,7 +352,7 @@ static int xmp_write(const char *path, const char *buf, size_t size,
 	syslog(LOG_ERR, "zhjwang...xmp_write...path:%s\n", path);
 	syslog(LOG_ERR, "zhjwang...xmp_write...size:%ld\n", size);
 	syslog(LOG_ERR, "=============================");
-	int res_write_cm = strcmp(path, "/home/zhjwang/test/1.wps");
+	int res_write_cm = strcmp(path, "/home/zhjwang/test/1.et");
 	if (res_write_cm == 0 && write_block_flag == 0)
 	{
 		write_block_flag = 1;
